Started Pseudocode
Started working through the steps that would be needed to make this program. At first I was going to throw all the menu creation in main, but then I decided to filter it out through separate functions. I will be trying to implement a way to create a singular menu builder helper function that can try to cut down on the repetitiveness of asking a user for some basic stuff to do basic multiplication. For the actual caluclations, I have a function in mind for that, but I would like a function to build each menu to make the code look easier and neater, more versatile, easier to maintain, and all of that.