Fin
Decided to not move forward on the complicated calculator. It was gonna take I think more time than I had, and the assignment is due today, so I have to have a finished project to present. In the future, I may change this to have a better implementation, but doubtful.